Split out tag-extractor to its own file.
